Fix Hugo relref references to work in md (#100)
This PR fixes the "ref" and "relref" links in the documentation so that
they work in Markdown and thus Github. This will mean the links will not
work in the Hugo documentation.

We need to figure out a way to get the links working in both but they
really should work in Github.

The links are now all absolute links in order to pass linkspector and to
work in md.

The linkspector tool fails on any link with a '#' in it, that is a link
to a tag in any document. The change to the .linkspector file is to
remove the '#' character and any following text, thus allowing those
links to pass linkspector.