Adding API Specs (#116)
So far it is an iframe reference to doc.crds.dev. 

In the long run it might be better to use the
[swaggerui](https://www.docsy.dev/docs/adding-content/shortcodes/#swaggerui)
or the
[redoc](https://www.docsy.dev/docs/adding-content/shortcodes/#redoc)
options of Docsy. For these two later options we would need to have the
Open API specs of the API-s somewhere. I have a feeling that at the
moment we have the API-s defined in Go code and not in Open API. Is this
correct @johnbelamaric ?

Signed-off-by: Gergely Csatari <gergely.csatari@nokia.com>