License update to CC-BY-4.0 (#120)
This PR is the result of the discussion on Nephio TSC channel regarding
the docs licensing. It was confirmed that CC-BY-4.0 license had to be
used and two things had to be changed:

1. Primary LICENSE file
2. SPDX identifier in code files

Quoting verbatim:
---Jenn Bonner (on 21st March 2024)---
Thread regarding the Docs repo incorrect license.
I've heard back from LF Legal. They've done a thorough review, and
recommend the following steps:

1. replace the Apache-2.0 license information in the Documentation repo
2. change all the SPDX identifiers for documentation int hat repo to
CC-BY-4.0 from Apache-2.0
3. Additionally, please have EasyCLA changed so that the documentation
repo is not included in EasyCLA. [DONE]
---

---------

Signed-off-by: Rahul Jadhav <nyrahul@gmail.com>