Add K8S_CONTEXT label documentation (#143)
This is contingent on, and related to
https://github.com/nephio-project/test-infra/pull/269 so will be
submitted as a draft for the time being. Documenting the additional
variable of K8S_CONTEXT introduced in the linked PR for deploying Nephio
components on non-kind clusters.

Note also that in the "Real K8s Cluster" installation instructions the
branch should be updated once
https://github.com/nephio-project/test-infra/pull/269 is merged and a
new tag is available.

Closes https://github.com/nephio-project/nephio/issues/577

---------

Co-authored-by: Liam Fallon <35595825+liamfallon@users.noreply.github.com>