Remove ubuntu20.04 support (#240)
Ubuntu 20.04 is now deprecated, so we should remove it from our support.

Apart from that, I added a small note that we have not tested on Ubuntu
24 at the moment. The porch and nephio controllers work fine on 24. The
issue is mostly with the containerlab router. Probably we need to update
its image.

I also updated the tag from v3.0.0 to `main` because in `main` we should
keep the tag `main`

Closing https://github.com/nephio-project/nephio/issues/923

Signed-off-by: Sagar Arora <sagar.arora@openairinterface.org>