Relref patch (#280)
The relref syntax used previously was using the old shortcode in hugo
and as such caused warnings at build stage these have been amended to
use the up to date syntax.
Secondly when using the correct relref syntax problems arise when hugo
parses the "&" symbol in folder names and seems to ignore/nullify this
symbol leading to failing build. rectified by renaming the folders and
replacing any instance of "&" with "and"