Adding Inspecting Packages Guide (#287)
Adding a guide to how to view porch Package Revisions and their
contents.

Most of these Guide sections are meant to take out the detailed use
cases out of the cli-guide into their particular use cases as the
cli-guide's main goal is to act as a "A manual/dictionary" for the
porchctl commands/arguments and not describe their use cases.